test(meta): mirror test tree to source layout
- tests/ only mirrored the package; evals/hooks/script tests sat flat at the root with ad-hoc names
- move src tests under tests/mcp_server_polarion/, split hook tests into claude_hooks/, script test into github_scripts/
