+++ tests/claude_hooks/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the tracked Claude Code PreToolUse hooks in `.claude/hooks/`."""
