+++ tests/github_scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the CI helper scripts in `.github/scripts/`."""
